--- packages/code-analyzer-regex-engine/src/rule.ts
@@ -1,8 +1,0 @@
-interface Rule {
-    regex: RegExp;
-    fileExtensions: string[];
-    message: string;
-    description: string;
-}
-
-export type RuleMap = { [ruleName: string]: Rule };
